// src/enum/errors/errors.enum.ts
export enum ErrorMessages {
  NUMBER = "Must include number",
  SMALL_LETTER = "Must include small letter",
  LARGE_LETTER = "Must include large letter",
  SYMBOL = "Must include symbol",
  INVALID_EMAIL = "Invalid email,please write correct email",
  USER_NOT_FOUND = "No account with that email was found",
  WRONG_PASSWORD = "Incorrect password,please write correct password",
  DEFAULT = "Email or password was incorrect",
  REQUIRED = "Field is required",
}